experimenting infinitly large number in examples/basic/misc.sim
