+++ examples/basic/listdemo.sim
@@ -40,7 +40,7 @@ for i = 100 to 0 step -1 {
 }
 
 @""+nl
-lib = loadDynamicLibrary("user32")
+lib = ["76D40000", "SIMPLE_DYNAMIC_LIBRARY_", 0]
 list = [0,"one",2,3,4,5,6,7,[9,8,"ade"],lib]
 @list @""
