+++ examples/basic/misc.sim
@@ -1,22 +1,80 @@
+call simple.debugging.Throw
 
-list = ["0","1","2","3","4","5","6"]
-li = [[[[[[[[[[[[1]]]]]]]]]]]]
+num1 = "384783478374837847738738723878971897897438974983748973894783974782738947387498374897238994723897489327847982374837487374837286473647673362347636487336874638764786278346372648327647863274673286487632764732864786378264763287647826476383264872687468634873646328764873444444444444477777777777777777777777777777777777777777777777777777777777777777777777777777777777777777777734237647364278463862483268774687647687468723648736847683274638764877368747633746826487648776873648736284763278468376487638764873264877632746723673687326486387268376783267326937829623876178678236786786726278368746238783264378637647326827687236832763876378647836728367823647328623784632987000000000000000000000000000000000000000384783478374837847738738723878971897897438974983748973894783974782738947387498374897238994723897489327847982374837487374837286473647673362347636487336874638764786278346372648327647863274673286487632764732864786378264763287647826476383264872687468634873646328764873444444444444477777777777777777777777777777777777777777777777777777777777777777777777777777777777777777777734237647364278463862483268774687647687468723648736847683274638764877368747633746826487648776873648736284763278468376487638764873264877632746723673687326486387268376783267326937829623876178678236786786726278368746238783264378637647326827687236832763876378647836728367823647328623784632987000000000000000000000000000000000000000384783478374837847738738723878971897897438974983748973894783974782738947387498374897238994723897489327847982374837487374837286473647673362347636487336874638764786278346372648327647863274673286487632764732864786378264763287647826476383264872687468634873646328764873444444444444477777777777777777777777777777777777777777777777777777777777777777777777777777777777777777777734237647364278463862483268774687647687468723648736847683274638764877368747633746826487648776873648736284763278468376487638764873264877632746723673687326486387268376783267326937829623876178678236786786726278368746238783264378637647326827687236832763876378647836728367823647328623784632987000000000000000000000000000000000000000384783478374837847738738723878971897897438974983748973894783974782738947387498374897238994723897489327847982374837487374837286473647673362347636487336874638764786278346372648327647863274673286487632764732864786378264763287647826476383264872687468634873646328764873444444444444477777777777777777777777777777777777777777777777777777777777777777777777777777777777777777777734237647364278463862483268774687647687468723648736847683274638764877368747633746826487648776873648736284763278468376487638764873264877632746723673687326486387268376783267326937829623876178678236786786726278368746238783264378637647326827687236832763876378647836728367823647328623784632987000000000000000000000000000000000000000384783478374837847738738723878971897897438974983748973894783974782738947387498374897238994723897489327847982374837487374837286473647673362347636487336874638764786278346372648327647863274673286487632764732864786378264763287647826476383264872687468634873646328764873444444444444477777777777777777777777777777777777777777777777777777777777777777777777777777777777777777777734237647364278463862483268774687647687468723648736847683274638764877368747633746826487648776873648736284763278468376487638764873264877632746723673687326486387268376783267326937829623876178678236786786726278368746238783264378637647326827687236832763876378647836728367823647328623784632987000000000000000000000000000000000000000384783478374837847738738723878971897897438974983748973894783974782738947387498374897238994723897489327847982374837487374837286473647673362347636487336874638764786278346372648327647863274673286487632764732864786378264763287647826476383264872687468634873646328764873444444444444477777777777777777777777777777777777777777777777777777777777777777777777777777777777777777777734237647364278463862483268774687647687468723648736847683274638764877368747633746826487648776873648736284763278468376487638764873264877632746723673687326486387268376783267326937829623876178678236786786726278368746238783264378637647326827687236832763876378647836728367823647328623784632987000000000000000000000000000000000000000"
+num2 = "12345123451234512345"
+num = new BigNum(num1)
+num = num + num2
+@num.value()
 
-#@lengthOf(list)
-#@list[0]
-#@list[lengthOf(list)-1]
+class BigNum 
 
-for x = 0 to lengthOf(list) - 1 {
-	display ""+x +" "
-}
-@""
-for x in "hello World" {
-	display ""+x +" "
-}
-@""
-for x in list {
-	display ""+x +" "
-}
+	n = ""
 
-@""
-@li[0][0]
+	block BigNum(n)
+		/*tempArray = ""
+		count = 0 
+		nl = lengthOf(n) - 1
+		for i = 0 to nl {
+			if count == 9
+				count = 0 
+				this.n + (0+tempArray) tempArray = "" 
+			end
+			tempArray += n[i]
+			if i < nl count++ else this.n + (0+tempArray) tempArray = ""  end
+		} */
+		this.n = n
+		
+	block operator(operator,invalue)
+		switch operator {
+			case "+"
+				value = "" 
+				sum = ""
+				if isString(invalue) or isNumber(invalue) {
+					value += invalue
+				}
+				l1 = lengthOf(n) - 1
+				l2 = lengthOf(value) - 1
+				carryOver = 0
+				k = 0
+				i = 0
+				j = l2 
+				for i = l1 to 0 step -1 {
+					tempSum = (0 + n[i] + value[j] + carryOver) % 10
+					carryOver = (0 + n[i] + value[j] + carryOver) / 10 
+					carryOver = 0+__left(""+carryOver,substr(""+carryOver, ".") - 1)
+					sum += tempSum
+					if j == 0 break end
+					j-- 
+				} 
+				if (l1 > l2) { 
+					while (i >= 0 ) {
+						tempSum = (0 + n[i] + carryOver) % 10 
+						carryOver = (0 + n[i] + carryOver) / 10 
+						carryOver = 0+__left(""+carryOver,substr(""+carryOver, ".") - 1)
+						sum += tempSum
+						i--
+					}
+				elif l1 < l2 
+					while (j >= 0) {
+						tempSum = (0 + value[j] + carryOver) % 10
+						carryOver = (0 + value[j--] + carryOver) / 10 
+						carryOver = 0+__left(""+carryOver,substr(""+carryOver, ".") - 1)
+						sum += tempSum
+						j--
+					}
+				else 
+					if carryOver > 0 sum += 0 + carryOver end
+				}
+				this.n = ""
+				lSum = lengthOf(sum) - 1
+				for i = lSum to 0 step -1 {
+					this.n += sum[i]
+				}
+		}
+		return this
+		
+	block value()
+		return n
+		
+#9879888888888888888444444444488888888888888888888888888888888888888
+#8798888888888888884444444444488888888888888888888888888888888888888
