+++ simple/sources/simple_misc.c
@@ -19,7 +19,6 @@ int SKIP_ERROR = 0 ;
 char *DEFAULT_FILE_NAME = "" ;
 char simple_file_initial_dir[SIMPLE_PATHSIZE];
 
-int list_index = 0 ;
 char* time_frame = "seconds" ;
 
 SIMPLE_API void print_line ( void )
