+++ simple/sources/simple_vmstrindex.c
@@ -55,12 +55,12 @@ void simple_vm_stsimple_index ( VM *vm, String *string, double nNum1 )
 {
 	char *newstr  ; 
 	/* We will push a pointer of the sub char to the stack */
-	if ( nNum1 < list_index || nNum1 > simple_string_size(string) ) {
+	if ( nNum1 < 1 || nNum1 > simple_string_size(string) ) {
             simple_vm_error(vm,SIMPLE_VM_ERROR_INDEXOUTOFRANGE);
             return ;
 	}
 	newstr = string->str ;
-	newstr = newstr + ((int) (nNum1 - list_index) ) ;
+	newstr = newstr + ((int) (nNum1 - 1) ) ;
 	SIMPLE_VM_STACK_PUSHPVALUE(newstr);
 	SIMPLE_VM_STACK_OBJTYPE = SIMPLE_OBJTYPE_SUBSTRING ;
 }
